requirements.txt

Django==3.2.7
django-tinymce==3.3.0
gunicorn==20.1.0
Pillow==8.3.2
psycopg2-binary==2.9.1
whitenoise==5.3.0
django-tinymce==3.3.0
django-cors-headers
django-brotli 